chore: moved osmosis swapper getTradeQuote into separate function
